backend/requirements.txt
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2023.11.17 ; python_version >= "3.10" and python_version < "4.0"
cffi==1.16.0 ; python_version >= "3.10" and python_version < "4.0" 
charset-normalizer==3.3.2 ; python_version >= "3.10" and python_version < "4.0" 
cryptography==42.0.1 ; python_version >= "3.10" and python_version < "4.0" 
django-appconf==1.0.6 ; python_version >= "3.10" and python_version < "4.0"
django-channels==0.7.0 ; python_version >= "3.10" and python_version < "4.0" 
django-cors-headers==4.3.1 ; python_version >= "3.10" and python_version < "4.0" 
django-cryptography==1.1 ; python_version >= "3.10" and python_version < "4.0"
django-sslserver==0.22 ; python_version >= "3.10" and python_version < "4.0"
django-webpack-loader==3.0.1 ; python_version >= "3.10" and python_version < "4.0" 
django==5.0.1 ; python_version >= "3.10" and python_version < "4.0" 
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0" 
idna==3.6 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pytz==2023.4 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0" 
requests==2.31.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0" 
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "4.0" 
typing-extensions==4.9.0 ; python_version >= "3.10" and python_version < "3.11" 
tzdata==2023.4 ; python_version >= "3.10" and python_version < "4.0" 
urllib3==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "4.0" 